added feature to select specific city states
created MultiNationSelectPopup to reuse for the city states
