+++ android/assets/jsons/translations/template.properties
@@ -336,8 +336,11 @@ Promotions =
 Load copied data = 
 Reset to defaults = 
 Select nations = 
+Select city states = 
 Random nations pool = 
+Random city states pool = 
 Blacklist random nations pool = 
+Blacklist random city states pool = 
 Are you sure you want to reset all game options to defaults? = 
 Start game! = 
 Map Options =
