+++ core/src/com/unciv/models/metadata/GameParameters.kt
@@ -29,8 +29,11 @@ class GameParameters : IsPartOfGameInfoSerialization { // Default values are the
     var numberOfCityStates = 6
 
     var enableRandomNationsPool = false
+    var enableRandomCityStatesPool = false
     var blacklistRandomNationsPool = false
+    var blacklistRandomCityStatesPool = false
     var randomNations = arrayListOf<Nation>()
+    var randomCityStates = arrayListOf<Nation>()
 
     var noCityRazing = false
     var noBarbarians = false
